Implement proper termination of remaining containers when instance is stopped
When draining a Kubernetes node with 'kubectl drain', DaemonSets are typically
not evicted from the node.

When stopping a Bosh instance though, the Bosh semantics require all processes
to terminate. So, after Kubelet is stopped and when ContainerD is still
running, we list all remaining containers at ContainerD level, and terminate
them. This preserves Bosh semantics, and stopped Bosh instance running
Kubernetes nodes won't run any unnecessary leftover process.

As an effect from this change, the kubelet is stopped during the 'drain'
script execution, which ensures that it is stopped before ContainerD.